Skip re-downloading content

+++ hdhr-dvr-jellyfin-plugin/SyncTask.cs
@@ -40,13 +40,17 @@ public class SyncTask(IConfigurationManager config, IHttpClientFactory httpClien
             } else if (episode.Item2.RecordEndTime == 0 || episode.Item2.RecordEndTime + 30 > DateTimeOffset.UtcNow.ToUnixTimeSeconds()) {
                 logger.LogInformation("Not downloading {SeriesName} episode {EpisodeNumber} because it is still recording", episode.Item1.Title, episode.Item2.EpisodeNumber);
             } else {
-                logger.LogInformation("Found {SeriesName} episode {EpisodeNumber} to download from {Url}", episode.Item1.Title, episode.Item2.EpisodeNumber, episode.Item2.PlayURL);
                 string seriesDir = Path.Join(dir, string.Concat(string.Format("{0} ({1})", episode.Item1.Title, episode.Item1.SeriesID).Split(Path.GetInvalidFileNameChars())));
                 _ = Directory.CreateDirectory(seriesDir);
                 string target = Path.Join(seriesDir, episode.Item2.Filename ?? string.Format("{0} ({1})", episode.Item2.Title, episode.Item2.EpisodeNumber));
-                await File.WriteAllTextAsync(target + ".storage.json", JsonSerializer.Serialize(episode.Item1), cancellationToken);
-                await File.WriteAllTextAsync(target + ".episode.json", JsonSerializer.Serialize(episode.Item2), cancellationToken);
-                await downloadManager.Add(episode.Item2.PlayURL, target, cancellationToken);
+                if (File.Exists(target)) {
+                    logger.LogInformation("Skipping download of {SeriesName} episode {EpisodeNumber} because it already exists at {Path}", episode.Item1.Title, episode.Item2.EpisodeNumber, target);
+                } else {
+                    logger.LogInformation("Found {SeriesName} episode {EpisodeNumber} to download from {Url}", episode.Item1.Title, episode.Item2.EpisodeNumber, episode.Item2.PlayURL);
+                    await File.WriteAllTextAsync(target + ".storage.json", JsonSerializer.Serialize(episode.Item1), cancellationToken);
+                    await File.WriteAllTextAsync(target + ".episode.json", JsonSerializer.Serialize(episode.Item2), cancellationToken);
+                    await downloadManager.Add(episode.Item2.PlayURL, target, cancellationToken);
+                }
             }
         }
         await downloadManager.Run(progress, cancellationToken);
